feat: move SMS mutation logic to model
